add font to menu

+++ src/config.rs
@@ -32,4 +32,4 @@ pub const HIT_SHOW_DELAY:u8 = 2;
 
 //font
 pub const FONT_SCALE_FAKTOR:u32 = 4;
-pub const FONT_Y_DISTANCE:i32 = 120/FONT_SCALE_FAKTOR as i32;
+pub const FONT_Y_DISTANCE:u32 = 120/FONT_SCALE_FAKTOR;
